+++ src/main/cliapp/src/constants/FilterFields.js
@@ -8,6 +8,10 @@ export const FIELD_SETS = Object.freeze({
     filterName: "abstractFilter",
     fields: ["abstract"],
   },
+  alleleFunctionalImpactsFieldSet: {
+    filterName: "alleleFunctionalImpactsFilter",
+    fields: ["alleleFunctionalImpacts.functionalImpacts.name", "alleleFunctionalImpacts.phenotypeTerm.curie", "alleleFunctionalImpacts.phenotypeTerm.name", "alleleFunctionalImpacts.phenotypeStatement", "alleleFunctionalImpacts.evidence.curie"],
+  },
   alleleNameFieldSet: {
     filterName: "alleleNameFilter",
     fields: ["alleleFullName.displayText", "alleleFullName.formatText"],
@@ -23,6 +27,10 @@ export const FIELD_SETS = Object.freeze({
   alleleSynonymsFieldSet: {
     filterName: "alleleSynonymsFilter",
     fields: ["alleleSynonyms.displayText", "alleleSynonyms.formatText"],
+  },
+  alleleGermlineTransmissionStatusFieldSet: {
+    filterName: "alleleGermlineTransmissionStatusFilter",
+    fields: ["alleleGermlineTransmissionStatus.germlineTransmissionStatus.name", "alleleGermlineTransmissionStatus.evidence.curie"]
   },
   alleleInheritanceModesFieldSet: {
     filterName: "alleleInheritanceModesFilter",
@@ -353,6 +361,8 @@ export const FIELD_SETS = Object.freeze({
 export const FILTER_CONFIGS = Object.freeze({
   abbreviationFilterConfig:                 { filterComponentType: 'input', fieldSets: [FIELD_SETS.abbreviationFieldSet] },
   abstractFilterConfig:                     { filterComponentType: "input", fieldSets: [FIELD_SETS.abstractFieldSet]},
+  alleleFunctionalImpactsFilterConfig:      { filterComponentType: "input", fieldSets: [FIELD_SETS.alleleFunctionalImpactsFieldSet] },
+  alleleGermlineTransmissionStatusFilterConfig: {filterComponentType: "input", fieldSets: [FIELD_SETS.alleleGermlineTransmissionStatusFieldSet]},
   alleleInheritanceModesFilterConfig:       { filterComponentType: "input", fieldSets: [FIELD_SETS.alleleInheritanceModesFieldSet] },
   alleleMutationFilterConfig:               { filterComponentType: "input", fieldSets: [FIELD_SETS.alleleMutationFieldSet] },
   alleleNameFilterConfig:                   { filterComponentType: "input", fieldSets: [FIELD_SETS.alleleNameFieldSet] },