+++ src/main/cliapp/src/containers/vocabularyPage/__tests__/VocabulariesTable.test.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { act } from "react-dom/test-utils";
 import { waitFor } from "@testing-library/react";
 import { renderWithClient } from '../../../tools/jest/utils';
 import { VocabulariesPage } from "../index";
@@ -17,38 +16,31 @@ describe("<VocabulariesPage />", () => {
 	});
 
 	it("Renders without crashing", async () => {
-		let result;
-		act(() => {
-			result = renderWithClient(<VocabulariesPage />);
-		});
-		
+		let result = await renderWithClient(<VocabulariesPage />);
+
 		await waitFor(() => {
 			expect(result);
 		});
 	});
 
 	it("Contains Correct Table Name", async () => {
-		let result;
-		act(() => {
-			result = renderWithClient(<VocabulariesPage />);
-		});
+		let result = await renderWithClient(<VocabulariesPage />);
 
 		const tableTitle = await result.findByText(/Vocabularies Table/i);
 		expect(tableTitle).toBeInTheDocument();
 	});
 
-	//skip this test for now, as it is not working
-	it.skip("The table contains correct data", async () => {
-		let result;
-		act(() => {
-			result = renderWithClient(<VocabulariesPage />);
-		});
-		
+	it("The table contains correct data", async () => {
+		let result = await renderWithClient(<VocabulariesPage />);
+
 		const nameTd = await result.findByText(/Name type/i);
+		const descriptionTd = await result.findByText(/Type of name represented by a name annotation/i);
+		const obsoleteTd = await result.findByText(/false/i);
 
 		await waitFor(() => {
 			expect(nameTd).toBeInTheDocument();
-
+			expect(descriptionTd).toBeInTheDocument();
+			expect(obsoleteTd).toBeInTheDocument();
 		});
 	});
 });